+++ phenopy/likelihood.py
@@ -5,7 +5,7 @@ from configparser import NoOptionError, NoSectionError
 import joblib
 # import lightgbm as lgb
 
-from phenopy import generate_annotated_hpo_network
+from phenopy.util import generate_annotated_hpo_network
 from phenopy.config import config, logger
 from phenopy.util import encode_phenotypes, read_phenotype_groups
