+++ tests/test_likelihood.py
@@ -3,8 +3,7 @@ import os
 import unittest
 
 from phenopy.config import config
-from phenopy import generate_annotated_hpo_network
-from phenopy.util import read_phenotype_groups
+from phenopy.util import read_phenotype_groups, generate_annotated_hpo_network
 from phenopy.likelihood import predict_likelihood_moldx
